database.js szétszedése + REST-nek megfelelő api-elnevezések
